能成功部署到服务器的版本(修改了pom) 所有model加了getter和setter 添加了CommunityController 将所有id修改为Interger

+++ src/main/java/com/gc/model/Team.java
@@ -1,7 +1,6 @@
 package com.gc.model;
 
 import javax.persistence.*;
-import java.sql.Date;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -14,7 +13,7 @@ public class Team {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private long id;
+    private Integer id;
 
     @Column(name = "title", nullable = false)
     private String title;
@@ -64,11 +63,11 @@ public class Team {
         this.updateTime = updateTime;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
